sched-Micro-optimize-the-smart-wake-affine-logic
Smart wake-affine is using node-size as the factor currently, but the
overhead
of the mask operation is high.

Thus, this patch introduce the 'sd_llc_size' percpu variable, which will
record
the highest cache-share domain size, and make it to be the new factor,
in order
to reduce the overhead and make it more reasonable.

Tested-by: Davidlohr Bueso <davidlohr.bueso@hp.com>
Tested-by: Michael Wang <wangyun@linux.vnet.ibm.com>
Signed-off-by: Peter Zijlstra <peterz@infradead.org>
Acked-by: Michael Wang <wangyun@linux.vnet.ibm.com>
Cc: Mike Galbraith <efault@gmx.de>
Link: http://lkml.kernel.org/r/51D5008E.6030102@linux.vnet.ibm.com
[ Tidied up the changelog. ]
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Signed-off-by: flar2 <asegaert@gmail.com>
Signed-off-by: Simarpreet Singh <simar@linux.com>
Signed-off-by: SirRGB <sirrgb@proton.me>